Refactor FormInput component
